requirements.txt
# Core LLM framework
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23

# OpenAI Python SDK
openai>=1.40.0

# Local vector database
chromadb==0.5.5

# Environment variable management
python-dotenv==1.0.1

# Web UI server
flask==3.0.3
